refactor: move EntityState type into Entity file
