+++ frontend/src/game/entity/types/index.ts
@@ -1,9 +1,8 @@
 import Direction from './direction';
 import StateConfig from './stateConfig';
-import State from './state';
 import EntityClass from './entityClass';
 import EntityId from './entityIds';
 
 export {
-  Direction, State, StateConfig, EntityId, EntityClass,
+  Direction, StateConfig, EntityId, EntityClass,
 };
